Added new provider exception. Refactored asset_downloader test. Moved create_asset in helper functions file.
