+++ tests/resources/helper_functions.py
@@ -15,6 +15,7 @@ from enforce_typing import enforce_types
 
 from ocean_lib.agreements.file_objects import FilesTypeFactory
 from ocean_lib.config import Config
+from ocean_lib.data_provider.data_service_provider import DataServiceProvider
 from ocean_lib.example_config import ExampleConfig
 from ocean_lib.models.data_token import DataToken
 from ocean_lib.models.erc20_token import ERC20Token
@@ -382,6 +383,46 @@ def get_provider_fees() -> Dict[str, Any]:
     }
 
     return provider_fee
+
+
+def create_asset(ocean, publisher, config, metadata=None):
+    """Helper function for asset creation based on ddo_sa_sample.json."""
+    erc20_data = ErcCreateData(
+        template_index=1,
+        strings=["Datatoken 1", "DT1"],
+        addresses=[
+            publisher.address,
+            publisher.address,
+            ZERO_ADDRESS,
+            get_address_of_type(config, "Ocean"),
+        ],
+        uints=[ocean.web3.toWei("0.5", "ether"), 0],
+        bytess=[b""],
+    )
+
+    if not metadata:
+        metadata = {
+            "created": "2020-11-15T12:27:48Z",
+            "updated": "2021-05-17T21:58:02Z",
+            "description": "Sample description",
+            "name": "Sample asset",
+            "type": "dataset",
+            "author": "OPF",
+            "license": "https://market.oceanprotocol.com/terms",
+        }
+    data_provider = DataServiceProvider
+    file1_dict = {"type": "url", "url": "https://url.com/file1.csv", "method": "GET"}
+    file1 = FilesTypeFactory(file1_dict)
+    encrypt_response = data_provider.encrypt(
+        [file1], "http://172.15.0.4:8030/api/services/encrypt"
+    )
+    encrypted_files = encrypt_response.content.decode("utf-8")
+
+    ddo = ocean.assets.create(
+        metadata, publisher, encrypted_files, erc20_tokens_data=[erc20_data]
+    )
+
+    return ddo
 
 
 def create_basics(config, web3, data_provider):
